[android] Create a BlePwoDiscoverer
This encapsulates the ble scann logic.  Unfortunately, because the
scanning for the fragment and the service are still different, we only
replace the scanning code in the NearbyBeaconsFragment.
